# api/openapi.yaml
openapi: 3.1.0
info:
  title: GLACIATION Metadata Service
  description: The service exposes API to work with Distributed Knowledge Graph
  license:
    name: MIT License
    url: https://github.com/glaciation-heu/glaciation-metadata-service/blob/main/LICENSE
  version: 1.5.6
paths:
  /api/v0/graph:
    patch:
      tags:
      - Graph
      summary: Update Graph
      description: Update Distributed Knowledge Graph
      operationId: update_graph_api_v0_graph_patch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Request body must be in JSON-LD format. It must be compatible
                with GLACIATION metadata upper ontology.
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: string
                title: Response Update Graph Api V0 Graph Patch
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - Graph
      summary: Search Graph
      description: Execute SPARQL search query and return a response in JSON format.
      operationId: search_graph_api_v0_graph_get
      parameters:
      - name: query
        in: query
        required: true
        schema:
          type: string
          description: SELECT query in SPARQL language. It must be compatible with
            GLACIATION metadata upper ontology.
          title: Query
        description: SELECT query in SPARQL language. It must be compatible with GLACIATION
          metadata upper ontology.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v0/graph/update:
    get:
      tags:
      - Graph
      summary: Perform Update Query
      description: Execute SPARQL update query and return a response.
      operationId: perform_update_query_api_v0_graph_update_get
      parameters:
      - name: query
        in: query
        required: true
        schema:
          type: string
          description: SELECT query in SPARQL language. It must be compatible with
            GLACIATION metadata upper ontology.
          title: Query
        description: SELECT query in SPARQL language. It must be compatible with GLACIATION
          metadata upper ontology.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: string
                title: Response Perform Update Query Api V0 Graph Update Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Graph
      summary: Perform Post Update Query
      description: Execute SPARQL update query and return a response.
      operationId: perform_post_update_query_api_v0_graph_update_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Update query in SPARQL language. It must be compatible
                with GLACIATION metadata upper ontology.
              title: Query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: string
                title: Response Perform Post Update Query Api V0 Graph Update Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v0/graph/compact:
    post:
      tags:
      - Graph
      summary: Perform Compaction
      operationId: perform_compaction_api_v0_graph_compact_post
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: string
                title: Response Perform Compaction Api V0 Graph Compact Post
  /metrics:
    get:
      tags:
      - Monitoring
      summary: Metrics
      description: Endpoint that serves Prometheus metrics.
      operationId: metrics_metrics_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ResponseHead:
      properties:
        vars:
          items:
            type: string
          type: array
          title: Vars
      type: object
      required:
      - vars
      title: ResponseHead
    ResponseResults:
      properties:
        bindings:
          items:
            type: object
          type: array
          title: Bindings
      type: object
      required:
      - bindings
      title: ResponseResults
    SearchResponse:
      properties:
        head:
          $ref: '#/components/schemas/ResponseHead'
        results:
          $ref: '#/components/schemas/ResponseResults'
      type: object
      required:
      - head
      - results
      title: SearchResponse
      example:
        head:
          vars:
          - sub
          - pred
          - obj
        results:
          bindings:
          - obj:
              type: uri
              value: http://data.kasabi.com/dataset/cheese/schema/Cheese
            pred:
              type: uri
              value: http://www.w3.org/1999/02/22-rdf-syntax-ns#type
            sub:
              type: uri
              value: http://data.kasabi.com/dataset/cheese/halloumi
          - obj:
              type: literal
              value: Halloumi
              xml:lang: el
            pred:
              type: uri
              value: http://www.w3.org/2000/01/rdf-schema#label
            sub:
              type: uri
              value: http://data.kasabi.com/dataset/cheese/halloumi
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError